backend/requirements.txt
Django==4.2
djangorestframework==3.15.1
ragflow-sdk==0.0.6
